keep a slice with the visible items to fix positioning bug
in windows, Hide() destroys a menuItem. this has problems when later we
try to Show() it and end up using the original menuID, for instance if
we did hide several elements in a row: if we insert the element in the
original position, it will end up inserted past some other elements that
were intended to be placed after the items that were hidden.

by keeping a slice where we insert and delete the id of the elements
that are shown, we can sort this slice and take the index of a given
element as the correct index in which we want to insert the menuItem
when calling Show() again.

- Resolves: #72